Fix startup hang and other cold start issues

+++ rootfs/etc/profile.d/fzf.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Install fzf shell completion when an interactive shell is enabled
 # This is a fix for: `/etc/bash_completion.d/fzf.sh: line 34: bind: warning: line editing not enabled`
-if [ -t 1 ] && ! [ -e '/etc/bash_completion.d/fzf.sh' ]; then
+if [ -t 1 ] && ! [ -e '/etc/bash_completion.d/fzf.sh' ] && [ -e '/usr/share/bash-completion/completions/fzf' ]; then
 	ln -s /usr/share/bash-completion/completions/fzf /etc/bash_completion.d/fzf.sh
 fi
